编写动态，评论，图片的授权策略

+++ app/Http/Controllers/CommentsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Comment;
+use App\Models\Status;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
@@ -24,10 +25,12 @@ class CommentsController extends Controller
             'content' => 'required|max:100'
         ]);
 
-        Auth::user()->comments()->create([
-            'content' => $request->input('content'),
-            'status_id' => $request->input('status')
-        ]);
+        $comment = new Comment;
+        $comment->content = $request->input('content');
+        $comment->status_id = $request->input('status');
+
+        $this->authorize('store', $comment);
+        Auth::user()->comments()->save($comment);
 
         session()->flash('success', '评论成功！');
         return redirect()->back();
@@ -36,6 +39,7 @@ class CommentsController extends Controller
     public function destroy($id)
     {
         $comment = Comment::findOrFail($id);
+        $this->authorize('destroy', $comment);
         $comment->delete();
         session()->flash('success', '删除评论成功');
         return redirect()->back();
